created a particel system

+++ src/crepe/Particle.cpp
@@ -1,12 +1,13 @@
 #include "Particle.hpp"
 
+
 Particle::Particle(float lifespan, Position position, Position velocity)
     : lifespan(lifespan), position(position), velocity(velocity), timeInLife(0.0f) {}
 
 void Particle::update(float deltaTime) {
     timeInLife += deltaTime;
-    position.x += static_cast<int>(velocity.x * deltaTime);
-    position.y += static_cast<int>(velocity.y * deltaTime);
+    position.x += velocity.x * deltaTime;
+    position.y += velocity.y * deltaTime;
 }
 
 bool Particle::isAlive() const {
